// src/app/sistema/administrativo/empresas/enums/regime-da-empresa.ts
export enum RegimeDaEmpresa {
  SIMPLES_NACIONAL = 'SIMPLES_NACIONAL',
  LUCRO_PRESUMIDO = 'LUCRO_PRESUMIDO',
  LUCRO_REAL = 'LUCRO_REAL',
  MEI = 'MEI',
  IMUNE = 'IMUNE',
  ISENTA = 'ISENTA',
  EMPRESARIO_INDIVIDUAL = 'EMPRESARIO_INDIVIDUAL',
  EIRELI = 'EIRELI',
  SLU = 'SLU',
  LTDA = 'LTDA',
  SA_ABERTA = 'SA_ABERTA',
  SA_FECHADA = 'SA_FECHADA',
  SOCIEDADE_SIMPLES = 'SOCIEDADE_SIMPLES',
  ASSOCIACAO = 'ASSOCIACAO',
  ONG = 'ONG',
  COOPERATIVA = 'COOPERATIVA',
  INTITUICAO_RELIGIOSA = 'INTITUICAO_RELIGIOSA',
  ORG_PUBLICO = 'ORG_PUBLICO',
  OUTROS = 'OUTROS',
}